General reliability improvements
I'll be honest this is just a "make the repo up to date" commit, I really need to be better at doing the git for this
thing. Problem is I'm almost always exploring and trying to figure out what on Earth is going on, so I'm never quite
sure when a commit is appropriate.